Compute hashes of object data so that we can detect nondeterministic task executions.