@pcmoritz 2016-12-06
Here are some of the performance implications:

Before:
```
In [11]: a = np.zeros(1000)

In [12]: %timeit ray.put(a)
1000 loops, best of 3: 986 µs per loop

In [5]: a = np.zeros(10000)

In [6]: %timeit ray.put(a)
100 loops, best of 3: 2.77 ms per loop

In [13]: a = np.zeros(100000)

In [14]: %timeit ray.put(a)
100 loops, best of 3: 14.9 ms per loop
```

After:
```
a = np.zeros(1000)
x = ray.put(a)
In [16]: %timeit ray.worker.global_worker.plasma_client.hash(x.id())
1000 loops, best of 3: 492 µs per loop

In [18]: %timeit ray.put(a)
1000 loops, best of 3: 736 µs per loop

a = np.zeros(10000)
In [20]: %timeit ray.put(a)
100 loops, best of 3: 2.11 ms per loop

In [21]: x = ray.put(a)

In [22]: %timeit ray.worker.global_worker.plasma_client.hash(x.id())
1000 loops, best of 3: 1.58 ms per loop

In [26]: %timeit ray.put(a)
100 loops, best of 3: 14.5 ms per loop

x = ray.put(a)

In [28]: %timeit ray.worker.global_worker.plasma_client.hash(x.id())
100 loops, best of 3: 12.6 ms per loop
```

Hitting the cache limit:

Before:
```
In [4]: a = np.zeros(10000000)
In [5]: %timeit ray.put(a)
10 loops, best of 3: 1.95 s per loop

```

After:
```
a = np.zeros(10000000)
In [36]: %timeit ray.put(a)
1 loop, best of 3: 1.33 s per loop
```

So the increase in time seems to be drowned by noise here. We should re-benchmark with https://github.com/ray-project/ray/pull/64.