@robertnishihara 2016-12-06
On Mac, I get a somewhat anywhere from a 5x to 15x slowdown for `ray.put`.