@pcmoritz 2016-12-06
Interesting, I'd really like to understand what is going on here. On linux if I do

```
--- a/src/plasma/plasma_client.c
+++ b/src/plasma/plasma_client.c
@@ -375,7 +375,7 @@ void plasma_seal(plasma_connection *conn, object_id object_i
   plasma_request req = plasma_make_request(object_id);
   CHECK(plasma_send_request(conn->store_conn, PLASMA_SEAL, &req) >= 0);
   if (conn->manager_conn >= 0) {
-    plasma_compute_object_hash(conn, object_id, req.digest);
+    /* plasma_compute_object_hash(conn, object_id, req.digest); */
     CHECK(plasma_send_request(conn->manager_conn, PLASMA_SEAL, &req) >= 0);
   }
 }
```
it seems to get slower :(

I have no idea why.