@robertnishihara 2016-12-07
After the switch to xxhash, I'm the slowdown is 1.5x to 2x slowdown for ray.put on Mac.