@pcmoritz 2016-12-09
merged in https://github.com/ray-project/ray/pull/103